Implemented ActiveJob for Mention.create_all
