+++ app/jobs/mentions/create_all_job.rb
@@ -1,0 +1,10 @@
+module Mentions
+  class CreateAllJob < ApplicationJob
+    queue_as :mentions_create_all
+
+    def perform(notifiable_id, notifiable_type)
+      notifiable = notifiable_type.constantize.find_by(id: notifiable_id)
+      MentionsCreateAllService.call(notifiable) if notifiable
+    end
+  end
+end
